Cambio sumadecinales

+++ produccion/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 # Create your models here.
 class Valor_referencia(models.Model):
-    valor = models.DecimalField(max_digits=10, decimal_places=2, default=0)
+    valor = models.DecimalField(max_digits=11, decimal_places=2, default=0)
 
     class Meta:
         verbose_name = 'Valor referencia'
@@ -11,14 +11,14 @@ class Valor_referencia(models.Model):
 
 
 class Perseo_produccion(models.Model):
-    pedido=models.CharField(max_length=50, default=0)
-    actividad=models.CharField(max_length=100, default=0)
-    instalador=models.CharField(max_length=200, default=0)
-    fecha=models.DateField(default=0)
-    codigo=models.CharField(max_length=50, default=0)
-    cantidad=models.DecimalField(decimal_places=3, max_digits=10)
-    valor=models.IntegerField(default=0)
-    total=models.IntegerField(default=0)
+    pedido= models.CharField(max_length=50, default=0)
+    actividad= models.CharField(max_length=100, default=0)
+    instalador= models.CharField(max_length=200, default=0)
+    fecha= models.DateField(default=0)
+    codigo= models.CharField(max_length=50, default=0)
+    cantidad= models.DecimalField(decimal_places=3, max_digits=10)
+    valor= models.DecimalField(max_digits=11, decimal_places=2, default=0.00)
+    total= models.DecimalField(max_digits=11, decimal_places=2, default=0.00)
     acta=models.CharField(max_length=50, default=0)
     
     class Meta:
@@ -38,12 +38,12 @@ class NovedadProduccion(models.Model):
 class Dia_dia(models.Model):
     instalador = models.CharField(max_length=200)
     fecha = models.DateField()
-    mano_obra= models.IntegerField(default=0)
-    materiales = models.IntegerField(default=0)
-    valor_referencia = models.IntegerField(default=0)
-    producido = models.IntegerField(default=0)
-    treinta_porciento = models.IntegerField(default=0)
-    por_persona = models.IntegerField(default=0)
+    mano_obra= models.DecimalField(max_digits=11, decimal_places=2, default=0.00)
+    materiales = models.DecimalField(max_digits=11, decimal_places=2, default=0.00)
+    valor_referencia = models.DecimalField(max_digits=11, decimal_places=2, default=0.00)
+    producido = models.DecimalField(max_digits=11, decimal_places=2, default=0.00)
+    treinta_porciento = models.DecimalField(max_digits=11, decimal_places=2, default=0.00)
+    por_persona = models.DecimalField(max_digits=11, decimal_places=2, default=0.00)
     
     class Meta:
